Added condition incidence counter.
Now a patient will not have a condition more than three times. After the
third incidence is treated, the module will terminate.

This is because with so many modules not terminating the speed of
generating a patient is much less (though there is way more data
generated per patient as a result)