Added ConditionEnd state.
Ends the current condition before either terminating the module or
attempting to restart the process.

Ideally the 'condtions.csv' file and the 'patient_conditions.csv' file
should have about the same number of conditions. This was not the case,
because the record which stores data for the conditions.csv file cannot
support more than one condition of the same type being active (which
makes sense). So marking a condition as ended solves this